+++ app/src/renderer/components/staking/LiValidator.vue
@@ -2,6 +2,7 @@
 .li-validator(:class='styles'): .li-validator__values
   .li-validator__value.name
     router-link(:to="{ name: 'validator', params: { validator: delegate.id }}")
+      span.validator-profile__status(v-bind:class="statusColor" v-tooltip.top="status")
       img.avatar(v-if="delegate.keybase" :src="delegate.keybase.avatarUrl" width="48" height="48")
       img.avatar(v-else src="~assets/images/validator-icon.svg" width="48" height="48")
       .vert
@@ -10,39 +11,30 @@
   .li-validator__value.your-votes
     span {{ yourVotes }}
   .li-validator__value.your-rewards
-    span {{ yourRewards }}
+    span n/a
   .li-validator__break: span
   .li-validator__value.percent_of_vote
-    span {{ delegate.percent_of_vote }}
+    span {{ delegate.percent_of_vote ? delegate.percent_of_vote : `n/a` }}
   .li-validator__value.uptime
     // add .green .yellow or .red class to this span to trigger inidication by color
     span {{ uptime }}
   .li-validator__value.commission
-    // add .green .yellow or .red class to this span to trigger inidication by color
     span {{ commission }}
   .li-validator__value.slashes
     // add .green .yellow or .red class to this span to trigger inidication by color
     span {{ slashes }}
-  template(v-if="!disabled")
-    .li-validator__value.checkbox(v-if="committedDelegations[delegate.id]")
-      i.material-icons lock
-    .li-validator__value.checkbox#remove-from-cart(v-else-if="inCart" @click='rm(delegate)')
-      i.material-icons check_box
-    .li-validator__value.checkbox#add-to-cart(v-else @click='add(delegate)')
-      i.material-icons check_box_outline_blank
-  template(v-else)
-    .li-validator__value
 </template>
 
 <script>
 import { mapGetters } from "vuex"
 import num from "scripts/num"
-import { shortAddress, calculateTokens } from "scripts/common"
+import { shortAddress, calculateTokens, ratToBigNumber } from "scripts/common"
 export default {
   name: "li-validator",
   props: ["delegate", "disabled"],
   computed: {
     ...mapGetters([
+      "lastHeader",
       "shoppingCart",
       "delegates",
       "config",
@@ -53,14 +45,37 @@ export default {
       return "n/a" //TODO: add slashes
     },
     commission() {
-      return "n/a" //TODO: add commission
+      return `${this.num.pretty(this.delegate.commission)}%`
     },
     uptime() {
-      return "n/a" //TODO: add real uptime
-    },
-    yourRewards() {
+      let rollingWindow = 10000 // param of slashing period
+      let info = this.delegate.signing_info
+      if (info) {
+        // uptime in the past 10k blocks
+        let uptimeRollingWindow = info.signed_blocks_counter / rollingWindow
+        return `${this.num.pretty(uptimeRollingWindow * 100)}%`
+      }
       return "n/a"
     },
+    // TODO uncomment when distribution is done
+    // yourRewards() {
+    //   if (
+    //     this.committedDelegations[this.delegate.id] &&
+    //     this.committedDelegations[this.delegate.id].isValidator &&
+    //     this.delegate.proposer_reward_pool > 0
+    //   ) {
+    //     let myShares = calculateShares(
+    //       this.delegate,
+    //       this.committedDelegations[this.delegate.id]
+    //     )
+    //     let shareRatio = myShares.div(parseValidatorShares(this.delegate))
+    //     let rewardsInShares = shareRatio.times(
+    //       this.delegate.proposer_reward_pool
+    //     )
+    //     let rewardsInTokens = calculateTokens(rewardsInShares)
+    //     return this.num.pretty(rewardsInTokens).toString()
+    //   } else return "0"
+    // },
     yourVotes() {
       return this.num.pretty(
         this.committedDelegations[this.delegate.id]
@@ -86,17 +101,43 @@ export default {
         : this.delegate.isValidator
           ? "Validator"
           : "Candidate"
-    }
-  },
-  data: () => ({ num, shortAddress }),
-  methods: {
-    add(delegate) {
-      this.$store.commit("addToCart", delegate)
     },
-    rm(delegate) {
-      this.$store.commit("removeFromCart", delegate.id)
+    powerRatio() {
+      return ratToBigNumber(this.delegate.tokens)
+        .div(this.delegates.globalPower)
+        .toNumber()
+    },
+    // TODO enable once we decide on limits
+    // powerRatioLevel() {
+    //   if (this.powerRatio < 0.01) return "green"
+    //   if (this.powerRatio < 0.03) return "yellow"
+    //   else if (this.powerRatio >= 0.03) return "red"
+    // },
+    status() {
+      // status: jailed
+      if (this.delegate.revoked)
+        return "This validator has been jailed and is not currently validating"
+
+      // status: candidate
+      if (parseFloat(this.delegate.voting_power) === 0)
+        return "This validator has declared candidacy but does not have enough voting power yet"
+
+      // status: validator
+      return "This validator is actively validating"
     }
-  }
+    // TODO enable once we decide on limits
+    // statusColor() {
+    //   // status: jailed
+    //   if (this.delegate.revoked) return "red"
+    //
+    //   // status: candidate
+    //   if (parseFloat(this.delegate.voting_power) === 0) return "yellow"
+    //
+    //   // status: validator
+    //   return "green"
+    // }
+  },
+  data: () => ({ num, shortAddress })
 }
 </script>
 
@@ -138,20 +179,20 @@ export default {
     text-align right
     padding 4px 4px
 
-    &.green
-      color var(--green)
-      background-color var(--green-fade-1)
-      border 1px solid var(--green-fade-2)
-
-    &.orange
-      color var(--orange)
-      background-color var(--orange-fade-1)
-      border 1px solid var(--orange-fade-2)
-
     &.red
-      color var(--red)
-      background-color var(--red-fade-1)
-      border 1px solid var(--red-fade-2)
+      background-color rgba(209, 2, 0, 0.15)
+      border solid 0.5px rgba(209, 2, 0, 0.25)
+      color #ff0200
+
+    &.yellow
+      background-color rgba(255, 149, 2, 0.15)
+      border solid 0.5px rgba(255, 149, 2, 0.25)
+      color #ff9502
+
+    &.green
+      background-color rgba(46, 164, 45, 0.15)
+      border solid 0.5px rgba(46, 164, 45, 0.25)
+      color #2ea42d
 
 .li-validator__break
   flex 0