+++ test/unit/specs/components/staking/PageValidator.spec.js
@@ -1,4 +1,5 @@
 import BigNumber from "bignumber.js"
+import Delegation from "renderer/vuex/modules/delegation"
 import ModalStake from "staking/ModalStake"
 import setup from "../../../helpers/vuex-setup"
 import PageValidator from "renderer/components/staking/PageValidator"
@@ -139,66 +140,91 @@ describe("PageValidator", () => {
     expect(wrapper.find("#validator-profile__self-bond").text()).toBe("1.00 %")
   })
 
-  it("switches color indicators", async () => {
+  it("should show the validator status", () => {
+    expect(wrapper.vm.status).toBe("This validator is actively validating")
+    // Jailed
     store.state.delegates.delegates = [
       Object.assign({}, delegate, {
-        commission: "0"
+        revoked: true
       })
     ]
     wrapper.update()
-    expect(wrapper.find("#validator-profile__commission").classes()).toContain(
-      "green"
+    expect(wrapper.vm.status).toBe(
+      "This validator has been jailed and is not currently validating"
     )
-
+    // Is not a validator
     store.state.delegates.delegates = [
       Object.assign({}, delegate, {
-        commission: "0.02"
+        voting_power: 0
       })
     ]
     wrapper.update()
-    expect(wrapper.find("#validator-profile__commission").classes()).toContain(
-      "yellow"
-    )
-
-    store.state.delegates.delegates = [
-      Object.assign({}, delegate, {
-        commission: "1"
-      })
-    ]
-    wrapper.update()
-    expect(wrapper.find("#validator-profile__commission").classes()).toContain(
-      "red"
-    )
-
-    store.state.delegates.globalPower = 1000
-    store.state.delegates.delegates = [
-      Object.assign({}, delegate, {
-        tokens: "1000"
-      })
-    ]
-    wrapper.update()
-    expect(wrapper.find("#validator-profile__power").classes()).toContain("red")
-
-    store.state.delegates.delegates = [
-      Object.assign({}, delegate, {
-        tokens: "10"
-      })
-    ]
-    wrapper.update()
-    expect(wrapper.find("#validator-profile__power").classes()).toContain(
-      "yellow"
-    )
-
-    store.state.delegates.delegates = [
-      Object.assign({}, delegate, {
-        tokens: "1"
-      })
-    ]
-    wrapper.update()
-    expect(wrapper.find("#validator-profile__power").classes()).toContain(
-      "green"
+    expect(wrapper.vm.status).toBe(
+      "This validator has declared candidacy but does not have enough voting power yet"
     )
   })
+
+  // TODO enable when we decide on limits are defined
+  // it("switches color indicators", async () => {
+  //   store.state.delegates.delegates = [
+  //     Object.assign({}, delegate, {
+  //       commission: "0"
+  //     })
+  //   ]
+  //   wrapper.update()
+  //   expect(wrapper.find("#validator-profile__commission").classes()).toContain(
+  //     "green"
+  //   )
+  //
+  //   store.state.delegates.delegates = [
+  //     Object.assign({}, delegate, {
+  //       commission: "0.02"
+  //     })
+  //   ]
+  //   wrapper.update()
+  //   expect(wrapper.find("#validator-profile__commission").classes()).toContain(
+  //     "yellow"
+  //   )
+  //
+  //   store.state.delegates.delegates = [
+  //     Object.assign({}, delegate, {
+  //       commission: "1"
+  //     })
+  //   ]
+  //   wrapper.update()
+  //   expect(wrapper.find("#validator-profile__commission").classes()).toContain(
+  //     "red"
+  //   )
+  //
+  //   store.state.delegates.globalPower = 1000
+  //   store.state.delegates.delegates = [
+  //     Object.assign({}, delegate, {
+  //       tokens: "1000"
+  //     })
+  //   ]
+  //   wrapper.update()
+  //   expect(wrapper.find("#validator-profile__power").classes()).toContain("red")
+  //
+  //   store.state.delegates.delegates = [
+  //     Object.assign({}, delegate, {
+  //       tokens: "10"
+  //     })
+  //   ]
+  //   wrapper.update()
+  //   expect(wrapper.find("#validator-profile__power").classes()).toContain(
+  //     "yellow"
+  //   )
+  //
+  //   store.state.delegates.delegates = [
+  //     Object.assign({}, delegate, {
+  //       tokens: "1"
+  //     })
+  //   ]
+  //   wrapper.update()
+  //   expect(wrapper.find("#validator-profile__power").classes()).toContain(
+  //     "green"
+  //   )
+  // })
 
   it("shows a validator as candidate if he has no voting_power", () => {
     store.state.delegates.delegates = [
@@ -208,9 +234,9 @@ describe("PageValidator", () => {
     ]
     wrapper.update()
     expect(wrapper.vm.status).toMatchSnapshot()
-    expect(wrapper.find(".validator-profile__status").classes()).toContain(
-      "yellow"
-    )
+    // expect(wrapper.find(".validator-profile__status").classes()).toContain(
+    //   "yellow"
+    // )
   })
 
   it("shows that a validator is revoked", () => {
@@ -221,9 +247,9 @@ describe("PageValidator", () => {
     ]
     wrapper.update()
     wrapper.vm.status = expect(wrapper.vm.status).toMatchSnapshot()
-    expect(wrapper.find(".validator-profile__status").classes()).toContain(
-      "red"
-    )
+    // expect(wrapper.find(".validator-profile__status").classes()).toContain(
+    //   "red"
+    // )
   })
 })
 
@@ -303,7 +329,7 @@ describe(`onStake`, () => {
         })
 
         expect($store.dispatch.mock.calls).toEqual([
-          ["submitDelegation", delegationTx]
+          [`submitDelegation`, { delegations: [{ atoms: 10, delegate }] }]
         ])
 
         expect($store.commit.mock.calls).toEqual([
@@ -341,7 +367,7 @@ describe(`onStake`, () => {
         })
 
         expect($store.dispatch.mock.calls).toEqual([
-          ["submitDelegation", delegationTx]
+          [`submitDelegation`, { delegations: [{ atoms: 10, delegate }] }]
         ])
 
         expect($store.commit.mock.calls).toEqual([
@@ -379,7 +405,7 @@ describe(`onStake`, () => {
         })
 
         expect($store.dispatch.mock.calls).toEqual([
-          [`submitDelegation`, delegationTx]
+          [`submitDelegation`, { delegations: [{ atoms: 10, delegate }] }]
         ])
 
         expect($store.commit.mock.calls).toEqual([
@@ -394,41 +420,106 @@ describe(`onStake`, () => {
       })
     })
 
-    // TODO update
-    // describe(`composition`, () => {
-    //   it(`delegation.submitDelegation`, async () => {
-    //     const delegation = Delegation({})
-    //
-    //     const dispatch = jest.fn((type, payload) => {
-    //       if (type === `submitDelegation`) {
-    //         delegation.actions[type]($store, payload)
-    //       }
-    //     })
-    //
-    //     const $store = {
-    //       commit: jest.fn(),
-    //       dispatch,
-    //       getters: getterValues,
-    //       rootState: getterValues,
-    //       state: {
-    //         committedDelegates: { "1a2b3c": 0 },
-    //         unbondingDelegations: {}
-    //       }
-    //     }
-    //
-    //     const {
-    //       vm: { submitDelegation }
-    //     } = mount(PageValidator, {
-    //       mocks: {
-    //         $route: { params: { validator: `1a2b3c`, delegator_shares: "19" } },
-    //         $store
-    //       }
-    //     })
-    //
-    //     await submitDelegation({ amount: 10 })
-    //
-    //     expect($store.dispatch.mock.calls).toEqual("hello")
-    //   })
-    // })
+    describe(`composition`, () => {
+      it(`delegation.submitDelegation`, async () => {
+        const delegation = Delegation({})
+
+        const dispatch = jest.fn((type, payload) => {
+          if (type === `submitDelegation`) {
+            delegation.actions[type]($store, payload)
+          }
+        })
+
+        const $store = {
+          commit: jest.fn(),
+          dispatch,
+          getters: getterValues,
+          rootState: getterValues,
+          state: {
+            committedDelegates: { "1a2b3c": 0 },
+            unbondingDelegations: {}
+          }
+        }
+
+        const {
+          vm: { submitDelegation }
+        } = mount(PageValidator, {
+          mocks: {
+            $route: { params: { validator: `1a2b3c` } },
+            $store
+          }
+        })
+
+        await submitDelegation({ amount: 10 })
+
+        expect($store.dispatch.mock.calls).toEqual([
+          [
+            "submitDelegation",
+            {
+              delegations: [
+                {
+                  atoms: 10,
+                  delegate: {
+                    bond_height: "0",
+                    bond_intra_tx_counter: 6,
+                    commission: "0.05",
+                    commission_change_rate: "0.01",
+                    commission_change_today: "0.005",
+                    commission_max: "0.1",
+                    delegator_shares: "19",
+                    description: {
+                      country: "DE",
+                      details: "Herr Schmidt",
+                      moniker: "herr_schmidt_revoked",
+                      website: "www.schmidt.de"
+                    },
+                    keybase: undefined,
+                    owner: "1a2b3c",
+                    prev_bonded_shares: "0",
+                    proposer_reward_pool: null,
+                    pub_key: {
+                      data: "dlN5SLqeT3LT9WsUK5iuVq1eLQV2Q1JQAuyN0VwSWK0=",
+                      type: "AC26791624DE60"
+                    },
+                    revoked: false,
+                    selfBond: 0.01,
+                    status: 2,
+                    tokens: "19",
+                    voting_power: "10"
+                  }
+                }
+              ]
+            }
+          ],
+          [
+            "sendTx",
+            {
+              begin_unbondings: [],
+              delegations: [
+                {
+                  delegation: { amount: "10", denom: "atom" },
+                  delegator_addr:
+                    "cosmosaccaddr15ky9du8a2wlstz6fpx3p4mqpjyrm5ctpesxxn9",
+                  validator_addr: "1a2b3c"
+                }
+              ],
+              to: "cosmosaccaddr15ky9du8a2wlstz6fpx3p4mqpjyrm5ctpesxxn9",
+              type: "updateDelegations"
+            }
+          ]
+        ])
+
+        expect($store.commit.mock.calls).toEqual([
+          ["setAtoms", 32],
+          [
+            "notify",
+            {
+              body: "You have successfully staked your atoms.",
+              title: "Successful Staking!"
+            }
+          ]
+        ])
+      })
+    })
   })
 })